+++ src/components/span.rs
@@ -295,6 +295,7 @@ mod tests {
 
     use crate::tui::style::Color;
     use crossterm::event::{KeyCode, KeyEvent};
+    use pretty_assertions::assert_eq;
 
     #[test]
     fn test_components_span() {